Add #283 `function-arity()` and incomplete tests
There appear to be some issues parsing, seemingly at the ANLTRv4 grammar
layer, with the following kinds of function expressions that are in the
spec for example test cases for this function in the XPath 3.1 spec.

`fn:function-arity(fn:substring#2)`
`let $initial := fn:substring(?, 1, 1) return fn:function-arity($initial) returns 1`

More investigation is needed and perhaps separate bugs must be filed.